+++ src/pmp/SurfaceMesh.h
@@ -12,8 +12,6 @@
 
 namespace pmp {
 
-class SurfaceMeshIO;
-
 //! \addtogroup core
 //!@{
 
@@ -1114,46 +1112,6 @@ public:
     SurfaceMesh& assign(const SurfaceMesh& rhs);
 
     //!@}
-    //! \name File IO
-    //!@{
-
-    //! \brief Read mesh from file \p filename controlled by \p flags
-    //! \details File extension determines file type. Supported formats and
-    //! vertex attributes (a=ASCII, b=binary):
-    //!
-    //! Format | ASCII | Binary | Normals | Colors | Texcoords
-    //! -------|-------|--------|---------|--------|----------
-    //! OFF    | yes   | yes    | a / b   | a      | a / b
-    //! OBJ    | yes   | no     | a       | no     | no
-    //! STL    | yes   | yes    | no      | no     | no
-    //! PLY    | yes   | yes    | no      | no     | no
-    //! PMP    | no    | yes    | no      | no     | no
-    //! XYZ    | yes   | no     | a       | no     | no
-    //! AGI    | yes   | no     | a       | a      | no
-    //!
-    //! In addition, the OBJ and PMP formats support reading per-halfedge
-    //! texture coordinates.
-    void read(const std::string& filename, const IOFlags& flags = IOFlags());
-
-    //! \brief Write mesh to file \p filename controlled by \p flags
-    //! \details File extension determines file type. Supported formats and
-    //! vertex attributes (a=ASCII, b=binary):
-    //!
-    //! Format | ASCII | Binary | Normals | Colors | Texcoords
-    //! -------|-------|--------|---------|--------|----------
-    //! OFF    | yes   | yes    | a       | a      | a
-    //! OBJ    | yes   | no     | a       | no     | no
-    //! STL    | yes   | no     | no      | no     | no
-    //! PLY    | yes   | yes    | no      | no     | no
-    //! PMP    | no    | yes    | no      | no     | no
-    //! XYZ    | yes   | no     | a       | no     | no
-    //!
-    //! In addition, the OBJ and PMP formats support writing per-halfedge
-    //! texture coordinates.
-    void write(const std::string& filename,
-               const IOFlags& flags = IOFlags()) const;
-
-    //!@}
     //! \name Add new elements by hand
     //!@{
 
@@ -2063,8 +2021,6 @@ private:
 
     // are there any deleted entities?
     inline bool has_garbage() const { return has_garbage_; }
-
-    friend SurfaceMeshIO; // code smell
 
     // property containers for each entity type and object
     PropertyContainer oprops_;