+++ src/pmp/io/SurfaceMeshIO.h
@@ -12,6 +12,41 @@
 
 namespace pmp {
 
+//! \brief Read into \p mesh from file \p filename controlled by \p flags
+//! \details File extension determines file type. Supported formats and
+//! vertex attributes (a=ASCII, b=binary):
+//!
+//! Format | ASCII | Binary | Normals | Colors | Texcoords
+//! -------|-------|--------|---------|--------|----------
+//! OFF    | yes   | yes    | a / b   | a      | a / b
+//! OBJ    | yes   | no     | a       | no     | no
+//! STL    | yes   | yes    | no      | no     | no
+//! PLY    | yes   | yes    | no      | no     | no
+//! XYZ    | yes   | no     | a       | no     | no
+//! AGI    | yes   | no     | a       | a      | no
+//!
+//! In addition, the OBJ and PMP formats support reading per-halfedge
+//! texture coordinates.
+void read(SurfaceMesh& mesh, const std::string& filename,
+          const IOFlags& flags = IOFlags());
+
+//! \brief Write \p mesh to file \p filename controlled by \p flags
+//! \details File extension determines file type. Supported formats and
+//! vertex attributes (a=ASCII, b=binary):
+//!
+//! Format | ASCII | Binary | Normals | Colors | Texcoords
+//! -------|-------|--------|---------|--------|----------
+//! OFF    | yes   | yes    | a       | a      | a
+//! OBJ    | yes   | no     | a       | no     | no
+//! STL    | yes   | no     | no      | no     | no
+//! PLY    | yes   | yes    | no      | no     | no
+//! XYZ    | yes   | no     | a       | no     | no
+//!
+//! In addition, the OBJ and PMP formats support writing per-halfedge
+//! texture coordinates.
+void write(const SurfaceMesh& mesh, const std::string& filename,
+           const IOFlags& flags = IOFlags());
+
 class SurfaceMeshIO
 {
 public:
@@ -29,7 +64,6 @@ private:
     void read_obj(SurfaceMesh& mesh);
     void read_stl(SurfaceMesh& mesh);
     void read_ply(SurfaceMesh& mesh);
-    void read_pmp(SurfaceMesh& mesh);
     void read_xyz(SurfaceMesh& mesh);
     void read_agi(SurfaceMesh& mesh);
 
@@ -38,7 +72,6 @@ private:
     void write_obj(const SurfaceMesh& mesh);
     void write_stl(const SurfaceMesh& mesh);
     void write_ply(const SurfaceMesh& mesh);
-    void write_pmp(const SurfaceMesh& mesh);
     void write_xyz(const SurfaceMesh& mesh);
 
     std::string filename_;